// frontend/src/metabase/embed/iframe-sdk/types/embed.ts
export type SdkIframeEmbedTagMessage =
  | SdkIframeEmbedTagIframeReadyMessage
  | SdkIframeEmbedTagRequestSessionTokenMessage
  | SdkIframeEmbedTagHandleLinkMessage
  | SdkIframeEmbedTagRequestGuestTokenRefreshMessage;

export type SdkIframeEmbedTagIframeReadyMessage = {
  type: "metabase.embed.iframeReady";
};
export type SdkIframeEmbedTagRequestSessionTokenMessage = {
  type: "metabase.embed.requestSessionToken";
};
export type SdkIframeEmbedTagHandleLinkMessage = {
  type: "metabase.embed.handleLink";
  data: { url: string; requestId: string };
};
export type SdkIframeEmbedTagRequestGuestTokenRefreshMessage = {
  type: "metabase.embed.requestGuestTokenRefresh";
  data: {
    expiredToken: string;
  };
};

/** Events that the sdk embed route listens for */
export type SdkIframeEmbedMessage =
  | SdkIframeEmbedSetSettingsMessage
  | SdkIframeEmbedSubmitSessionTokenMessage
  | SdkIframeEmbedReportAuthenticationError
  | SdkIframeEmbedReportAnalytics
  | SdkIframeEmbedHandleLinkResponse
  | SdkIframeEmbedSubmitRefreshedGuestTokenMessage;

export type SdkIframeEmbedSetSettingsMessage = {
  type: "metabase.embed.setSettings";
  data: SdkIframeEmbedSettings;
};
export type SdkIframeEmbedSubmitSessionTokenMessage = {
  type: "metabase.embed.submitSessionToken";
  data: {
    authMethod: MetabaseAuthMethod;
    sessionToken: MetabaseEmbeddingSessionToken;
  };
};
export type SdkIframeEmbedReportAuthenticationError = {
  type: "metabase.embed.reportAuthenticationError";
  data: {
    error: MetabaseError<MetabaseErrorCode, unknown>;
  };
};
export type SdkIframeEmbedReportAnalytics = {
  type: "metabase.embed.reportAnalytics";
  data: {
    usageAnalytics: EmbeddedAnalyticsJsEventSchema;
    embedHostUrl: string;
  };
};
export type SdkIframeEmbedHandleLinkResponse = {
  type: "metabase.embed.handleLinkResponse";
  data: { requestId: string; handled: boolean };
};
export type SdkIframeEmbedSubmitRefreshedGuestTokenMessage = {
  type: "metabase.embed.submitRefreshedGuestToken";
  data: {
    guestToken: string;
  };
};
// ...
export type SdkIframeEmbedSettings = Omit<
  SdkIframeEmbedBaseSettings,
  "fetchRequestToken"
> &
  SdkIframeEmbedTemplateSettings;